Several fixes to previously added features.
1. Add userspace size to track the size of the generated user space,
this is used for implementation MALLOC of userspace.
2. Add PEST rule for type qualifiers (e.g., const, mut). integration TBC
3. New function set_environment_variable_count,
this is used to count total env_vars from implementer cfg,
not those used in program, as this is the proper stride
that should be used. Old one was incorrect.
4. Add size method to PrimitiveDataType to return memory size of
 each primitive for use in ADD_P
5. Made ENVIRONMENTVARIABLE a DataType, this is to allow for
ENV VARS to be values, pointers, arrays all at once and still satisfy
 semantic analysis.
6. Add a new new built ins for memory/pointers.
These should only be used with expert knowledge but are exposed
 temporarily until arrays are completely finished (including all cases).
7. Add size parameter to BarracudaIR::Array.
Propagates the size of the array through to generation.
8. Temporarily make all variable/parameter types MUTABLE.
This is a fix until types/qualifiers are implemented properly for arrays
9. add semantic analysis for assigning to CONST, causes a panic.
Strictly defining non-assignment to CONST values.